update item obj name
