+++ Editor/FeatureSpawner.cs
@@ -31,11 +31,11 @@ namespace AlturaNFT.Editor
             Selection.activeGameObject= new GameObject(AlturaConstants.FeatureName_GetUsers).AddComponent<GetUsers>().gameObject;
         }
         
-        [MenuItem(AlturaConstants.BaseFeatureSpawnerMenu + AlturaConstants.FeatureName_NFTs_OfAccount)]
-        [MenuItem(GameObjMenu + AlturaConstants.FeatureName_NFTs_OfAccount)]
-       static void Spawn_NFtsOfAccount()
+        [MenuItem(AlturaConstants.BaseFeatureSpawnerMenu + AlturaConstants.FeatureName_GetUsersItems)]
+        [MenuItem(GameObjMenu + AlturaConstants.FeatureName_GetUsersItems)]
+       static void Spawn_GetUsersItems()
         {
-           Selection.activeGameObject= new GameObject(AlturaConstants.FeatureName_NFTs_OfAccount).AddComponent<NFTs_OwnedByAnAccount>().gameObject;
+           Selection.activeGameObject= new GameObject(AlturaConstants.FeatureName_GetUsersItems).AddComponent<GetUsersItems>().gameObject;
         }
         
         [MenuItem(AlturaConstants.BaseFeatureSpawnerMenu + AlturaConstants.FeatureName_AuthenticateUser)]
@@ -52,11 +52,11 @@ namespace AlturaNFT.Editor
             Selection.activeGameObject= new GameObject(AlturaConstants.FeatureName_ConnectUserWallet).AddComponent<ConnectPlayerWallet>().gameObject;
         }
         
-        [MenuItem(AlturaConstants.BaseFeatureSpawnerMenu + AlturaConstants.FeatureName_Txn_Account)]
-        [MenuItem(GameObjMenu + AlturaConstants.FeatureName_Txn_Account)]
-        static void Spawn_Txn_Account()
+        [MenuItem(AlturaConstants.BaseFeatureSpawnerMenu + AlturaConstants.FeatureName_GetCollections)]
+        [MenuItem(GameObjMenu + AlturaConstants.FeatureName_GetCollections)]
+        static void Spawn_GetCollections()
         {
-            Selection.activeGameObject= new GameObject(AlturaConstants.FeatureName_Txn_Account).AddComponent<Txn_Account>().gameObject;
+            Selection.activeGameObject= new GameObject(AlturaConstants.FeatureName_GetCollections).AddComponent<GetCollections>().gameObject;
         }
         
         [MenuItem(AlturaConstants.BaseFeatureSpawnerMenu + AlturaConstants.FeatureName_Txn_NFT)]
@@ -66,11 +66,11 @@ namespace AlturaNFT.Editor
             Selection.activeGameObject= new GameObject(AlturaConstants.FeatureName_Txn_NFT).AddComponent<GetItems>().gameObject;
         }
         
-        [MenuItem(AlturaConstants.BaseFeatureSpawnerMenu + AlturaConstants.FeatureName_Txn_Collection)]
-        [MenuItem(GameObjMenu + AlturaConstants.FeatureName_Txn_Collection)]
-        static void Spawn_Txn_Collection()
+        [MenuItem(AlturaConstants.BaseFeatureSpawnerMenu + AlturaConstants.FeatureName_GetCollection)]
+        [MenuItem(GameObjMenu + AlturaConstants.FeatureName_GetCollection)]
+        static void Spawn_GetCollection()
         {
-            Selection.activeGameObject= new GameObject(AlturaConstants.FeatureName_Txn_Collection).AddComponent<Txn_Collection>().gameObject;
+            Selection.activeGameObject= new GameObject(AlturaConstants.FeatureName_GetCollection).AddComponent<GetCollection>().gameObject;
         }
 
         [MenuItem(AlturaConstants.BaseFeatureSpawnerMenu + AlturaConstants.FeatureName_Transfer)]
