+++ Editor/GetCollection_Editor.cs
@@ -5,13 +5,13 @@ namespace AlturaNFT.Editor
     using UnityEditor;
     using Internal;
 
-    [CustomEditor(typeof(Txn_Account))]
-    public class Txn_Account_Editor : Editor
+    [CustomEditor(typeof(GetCollection))]
+    public class GetCollection_Editor : Editor
     {
         public override void OnInspectorGUI()
         {
             
-            Txn_Account myScript = (Txn_Account)target;
+            GetCollection myScript = (GetCollection)target;
             
             
             Texture banner = Resources.Load<Texture>("c_nftdata_details");
@@ -19,15 +19,14 @@ namespace AlturaNFT.Editor
             GUILayout.Box(banner);
             GUILayout.EndHorizontal();
 
-            if (GUILayout.Button("Get Account NFT Transactions", GUILayout.Height(45)))
+            if (GUILayout.Button("GET Contract/Collection Transactions", GUILayout.Height(45)))
             {
                 AlturaUser.SetFromEditorWin();
                 myScript.Run();
             }
-        
 
             if(GUILayout.Button("View Documentation", GUILayout.Height(25)))
-                Application.OpenURL(AlturaConstants.Docs_Txns_Account);
+                Application.OpenURL(AlturaConstants.Docs_GetCollection);
             DrawDefaultInspector();
         }
     }
