+++ Runtime/GetCollections.cs
@@ -1,35 +1,45 @@
-using System;
 using System.Collections;
-using System.Collections.Generic;
 using Newtonsoft.Json;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.Networking;
-using UnityEngine.Serialization;
 
 namespace AlturaNFT  
 { using Internal;
     
-
-    [AddComponentMenu(AlturaConstants.BaseComponentMenu+AlturaConstants.FeatureName_NFTs_OfAccount)]
+    /// <summary>
+    /// Get Many Collections
+    /// </summary>
+    [AddComponentMenu(AlturaConstants.BaseComponentMenu+AlturaConstants.FeatureName_GetCollections)]
     [ExecuteAlways]
-    [HelpURL(AlturaConstants.NFTs_OfAccount)]
-    public class NFTs_OwnedByAnAccount : MonoBehaviour
+    [HelpURL(AlturaConstants.Docs_GetCollections)]
+     public class GetCollections : MonoBehaviour
     {
 
         #region Parameter Defines
-
-            
+              
             [SerializeField]
-            private string address = "Input Account Address To Fetch NFT's from";
-            
+            private string _perPage = "Input How much pqges you want to get";
+            [SerializeField]
+            private string _page = "Input Which page you want to get";
+            [SerializeField]
+            private string _sortBy = "Input Sort By = name";
+            [SerializeField]
+            private string _sortDir = "Input Asc or Desc";
+
+            [Header("Optional: Filter and fetch items with specified property")]
+
+            [SerializeField]
+            [Tooltip("Filter from a documents by any properties")]
+            string _isVerified;
+            private string RequestUriInit = "https://api.alturanft.com/api/v2/collection";
             private string WEB_URL;
             private string _apiKey;
             private bool destroyAtEnd = false;
 
 
             private UnityAction<string> OnErrorAction;
-            private UnityAction<Items_model> OnCompleteAction;
+            private UnityAction<Collection_model> OnCompleteAction;
             
             [Space(20)]
             //[Header("Called After Successful API call")]
@@ -38,12 +48,13 @@ namespace AlturaNFT
             public UnityEvent afterError;
 
             [Header("Run Component when this Game Object is Set Active")]
-            [SerializeField] private bool onEnable = false;
+            [SerializeField] private bool onEnable = true;
             public bool debugErrorLog = true;
-            public bool debugLogRawApiResponse = false;
+            public bool debugLogRawApiResponse = true;
             
             [Header("Gets filled with data and can be referenced:")]
-            public Items_model item;
+            public Collection_model collectionModel;
+
 
         #endregion
 
@@ -70,28 +81,50 @@ namespace AlturaNFT
         /// Initialize creates a gameobject and assings this script as a component. This must be called if you are not refrencing the script any other way and it doesn't already exists in the scene.
         /// </summary>
         /// <param name="destroyAtEnd"> Optional bool parameter can set to false to avoid Spawned GameObject being destroyed after the Api process is complete. </param>
-        public static NFTs_OwnedByAnAccount Initialize(bool destroyAtEnd = true)
+        public static GetCollections Initialize(bool destroyAtEnd = true)
             {
-                var _this = new GameObject("NFTs Of Account").AddComponent<NFTs_OwnedByAnAccount>();
+                var _this = new GameObject(AlturaConstants.FeatureName_GetCollections).AddComponent<GetCollections>();
                 _this.destroyAtEnd = destroyAtEnd;
                 _this.onEnable = false;
                 _this.debugErrorLog = false;
                 return _this;
             }
-            
-            /// <summary>
-            /// Set Account Address to retrieve NFTs from as string
-            /// </summary>
-            /// <param name="account_address"> as string.</param>
-            public NFTs_OwnedByAnAccount SetAddress(string account_address)
+
+        /// <summary>
+        /// Set Parameters to retrieve User From .
+        /// </summary>
+        /// <param name="perPage"> amount of pages to query</param>
+        /// <param name="page"> page to query</param>
+        /// <param name="sortBy"> sort by field</param>
+        /// <param name="sortDir"> sort direction</param>
+        public GetCollections SetParameters(string perPage = "20", string page = "1", string sortBy = "name", string sortDir = "asc")
             {
-                this.address = account_address;
+
+                if(perPage!=null)
+                    this._perPage = perPage;
+                if(page!=null)
+                    this._page = page;
+                if(sortBy!=null)
+                    this._sortBy = sortBy;
+                if(sortDir!=null)
+                    this._sortDir = sortDir;
+     
+
                 return this;
             }
-  
+
+            /// <summary>
+            /// Set Filter by to return NFTs only from the given contract address/collection. 
+            /// </summary>
+            ///<param name="name"> as string.</param>
+            public GetCollections AlturaOptions(string isVerified)
+            {
+                this._isVerified = isVerified;
+                return this;
+            }
 
 
-            public NFTs_OwnedByAnAccount OnComplete(UnityAction<Items_model> action)
+            public GetCollections OnComplete(UnityAction<Collection_model> action)
             {
                 this.OnCompleteAction = action;
                 return this;
@@ -102,7 +135,7 @@ namespace AlturaNFT
             /// </summary>
             /// <param name="UnityAction action"> string.</param>
             /// <returns> Information on Error as string text.</returns>
-            public NFTs_OwnedByAnAccount OnError(UnityAction<string> action)
+            public GetCollections OnError(UnityAction<string> action)
             {
                 this.OnErrorAction = action;
                 return this;
@@ -115,29 +148,21 @@ namespace AlturaNFT
             /// <summary>
             /// Runs the Api call and fills the corresponding model in the component on success.
             /// </summary>
-            public Items_model Run()
+            public Collection_model Run()
             {
                 WEB_URL = BuildUrl();
                 StopAllCoroutines();
                 StartCoroutine(CallAPIProcess());
-                return item;
+                return collectionModel;
             }
 
             string BuildUrl()
             {
 
-                    WEB_URL = "https://api.alturanft.com/api/v2/user/items/" + address;
-           
-                
-                if (debugErrorLog)
-                {
-                    var s = "Querying NFTs owned of Account: " + address + " on " ;
-                    if (item.item.collectionAddress != "")
-                        s += " / Filter from collection: " + item.item.collectionAddress;
-                    Debug.Log(s);
+                    WEB_URL = RequestUriInit + "?perPage=" + _perPage + "&page=" + _page + "&sortBy=" + _sortBy + "&sortDir=" + _sortDir + "&isVerified=" + _isVerified;
+                    if(debugErrorLog)
+                        Debug.Log("Querying Details of User: " );
 
-                }
-                
                 return WEB_URL;
             }
             
@@ -146,8 +171,9 @@ namespace AlturaNFT
                 //Make request
                 UnityWebRequest request = UnityWebRequest.Get(WEB_URL);
                 request.SetRequestHeader("Content-Type", "application/json");
-                request.SetRequestHeader("Authorization", _apiKey);
                 request.SetRequestHeader("source", AlturaUser.GetSource());
+                request.SetRequestHeader("Authorization", _apiKey);
+                
 
                 {
                     yield return request.SendWebRequest();
@@ -158,34 +184,32 @@ namespace AlturaNFT
 
                     if (request.error != null)
                     {
-                        item = null;
                         if(OnErrorAction!=null)
                             OnErrorAction($"Null data. Response code: {request.responseCode}. Result {jsonResult}");
                         if(debugErrorLog)
-                            Debug.Log($" (⊙.◎) Null data. Response code: {request.responseCode}. Result {jsonResult}");
+                            Debug.Log($"Null data. Response code: {request.responseCode}. Result {jsonResult}");
                         if(afterError!=null)
                             afterError.Invoke();
+                        collectionModel = null;
                         //yield break;
                     }
                     else
                     {
-                        //Fill Data Model from recieved class
-                        item = JsonConvert.DeserializeObject<Items_model>(
+                        collectionModel = JsonConvert.DeserializeObject<Collection_model>(
                             jsonResult,
                             new JsonSerializerSettings
                             {
-                                NullValueHandling = NullValueHandling.Ignore,
-                                MissingMemberHandling = MissingMemberHandling.Ignore
+                            NullValueHandling = NullValueHandling.Ignore,
+                            MissingMemberHandling = MissingMemberHandling.Ignore
                             });
                         
                         if(OnCompleteAction!=null)
-                            OnCompleteAction.Invoke(item);
+                            OnCompleteAction.Invoke(collectionModel);
                         
                         if(afterSuccess!=null)
                             afterSuccess.Invoke();
                         
-                        if(debugErrorLog)
-                            Debug.Log("Success , view NFTs model" );
+                            Debug.Log($"Get Many Collection Models" );
                     }
                 }
                 request.Dispose();
