+++ Runtime/GetUsersItems.cs
@@ -1,45 +1,35 @@
+using System;
 using System.Collections;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.Networking;
+using UnityEngine.Serialization;
 
 namespace AlturaNFT  
 { using Internal;
     
-    /// <summary>
-    /// Get NFT Transactions of an account
-    /// </summary>
-    [AddComponentMenu(AlturaConstants.BaseComponentMenu+AlturaConstants.FeatureName_Txn_Account)]
+
+    [AddComponentMenu(AlturaConstants.BaseComponentMenu+AlturaConstants.FeatureName_GetUsersItems)]
     [ExecuteAlways]
-    [HelpURL(AlturaConstants.Docs_Txns_Account)]
-     public class Txn_Account : MonoBehaviour
+    [HelpURL(AlturaConstants.GetUsersItems)]
+    public class GetUsersItems : MonoBehaviour
     {
 
         #region Parameter Defines
-              
-            [SerializeField]
-            private string _perPage = "Input How much pqges you want to get";
-            [SerializeField]
-            private string _page = "Input Which page you want to get";
-            [SerializeField]
-            private string _sortBy = "Input Sort By = name";
-            [SerializeField]
-            private string _sortDir = "Input Asc or Desc";
 
-            [Header("Optional: Filter and fetch items with specified property")]
-
+            
             [SerializeField]
-            [Tooltip("Filter from a documents by any properties")]
-            string _isVerified;
-            private string RequestUriInit = "https://api.alturanft.com/api/v2/collection";
+            private string address = "Input Account Address To Fetch NFT's from";
+            
             private string WEB_URL;
             private string _apiKey;
             private bool destroyAtEnd = false;
 
 
             private UnityAction<string> OnErrorAction;
-            private UnityAction<Collection_model> OnCompleteAction;
+            private UnityAction<Items_model> OnCompleteAction;
             
             [Space(20)]
             //[Header("Called After Successful API call")]
@@ -48,13 +38,12 @@ namespace AlturaNFT
             public UnityEvent afterError;
 
             [Header("Run Component when this Game Object is Set Active")]
-            [SerializeField] private bool onEnable = true;
+            [SerializeField] private bool onEnable = false;
             public bool debugErrorLog = true;
-            public bool debugLogRawApiResponse = true;
+            public bool debugLogRawApiResponse = false;
             
             [Header("Gets filled with data and can be referenced:")]
-            public Collection_model collectionModel;
-
+            public Items_model item;
 
         #endregion
 
@@ -81,50 +70,28 @@ namespace AlturaNFT
         /// Initialize creates a gameobject and assings this script as a component. This must be called if you are not refrencing the script any other way and it doesn't already exists in the scene.
         /// </summary>
         /// <param name="destroyAtEnd"> Optional bool parameter can set to false to avoid Spawned GameObject being destroyed after the Api process is complete. </param>
-        public static Txn_Account Initialize(bool destroyAtEnd = true)
+        public static GetUsersItems Initialize(bool destroyAtEnd = true)
             {
-                var _this = new GameObject(AlturaConstants.FeatureName_Txn_Account).AddComponent<Txn_Account>();
+                var _this = new GameObject("NFTs Of Account").AddComponent<GetUsersItems>();
                 _this.destroyAtEnd = destroyAtEnd;
                 _this.onEnable = false;
                 _this.debugErrorLog = false;
                 return _this;
             }
-
-        /// <summary>
-        /// Set Parameters to retrieve User From .
-        /// </summary>
-        /// <param name="perPage"> amount of pages to query</param>
-        /// <param name="page"> page to query</param>
-        /// <param name="sortBy"> sort by field</param>
-        /// <param name="sortDir"> sort direction</param>
-        public Txn_Account SetParameters(string perPage = "20", string page = "1", string sortBy = "name", string sortDir = "asc")
-            {
-
-                if(perPage!=null)
-                    this._perPage = perPage;
-                if(page!=null)
-                    this._page = page;
-                if(sortBy!=null)
-                    this._sortBy = sortBy;
-                if(sortDir!=null)
-                    this._sortDir = sortDir;
-     
-
-                return this;
-            }
-
+            
             /// <summary>
-            /// Set Filter by to return NFTs only from the given contract address/collection. 
+            /// Set Account Address to retrieve NFTs from as string
             /// </summary>
-            ///<param name="name"> as string.</param>
-            public Txn_Account AlturaOptions(string isVerified)
+            /// <param name="account_address"> as string.</param>
+            public GetUsersItems SetAddress(string account_address)
             {
-                this._isVerified = isVerified;
+                this.address = account_address;
                 return this;
             }
+  
 
 
-            public Txn_Account OnComplete(UnityAction<Collection_model> action)
+            public GetUsersItems OnComplete(UnityAction<Items_model> action)
             {
                 this.OnCompleteAction = action;
                 return this;
@@ -135,7 +102,7 @@ namespace AlturaNFT
             /// </summary>
             /// <param name="UnityAction action"> string.</param>
             /// <returns> Information on Error as string text.</returns>
-            public Txn_Account OnError(UnityAction<string> action)
+            public GetUsersItems OnError(UnityAction<string> action)
             {
                 this.OnErrorAction = action;
                 return this;
@@ -148,21 +115,29 @@ namespace AlturaNFT
             /// <summary>
             /// Runs the Api call and fills the corresponding model in the component on success.
             /// </summary>
-            public Collection_model Run()
+            public Items_model Run()
             {
                 WEB_URL = BuildUrl();
                 StopAllCoroutines();
                 StartCoroutine(CallAPIProcess());
-                return collectionModel;
+                return item;
             }
 
             string BuildUrl()
             {
 
-                    WEB_URL = RequestUriInit + "?perPage=" + _perPage + "&page=" + _page + "&sortBy=" + _sortBy + "&sortDir=" + _sortDir + "&isVerified=" + _isVerified;
-                    if(debugErrorLog)
-                        Debug.Log("Querying Details of User: " );
+                    WEB_URL = "https://api.alturanft.com/api/v2/user/items/" + address;
+           
+                
+                if (debugErrorLog)
+                {
+                    var s = "Querying NFTs owned of Account: " + address + " on " ;
+                    if (item.item.collectionAddress != "")
+                        s += " / Filter from collection: " + item.item.collectionAddress;
+                    Debug.Log(s);
 
+                }
+                
                 return WEB_URL;
             }
             
@@ -171,9 +146,8 @@ namespace AlturaNFT
                 //Make request
                 UnityWebRequest request = UnityWebRequest.Get(WEB_URL);
                 request.SetRequestHeader("Content-Type", "application/json");
-                request.SetRequestHeader("source", AlturaUser.GetSource());
                 request.SetRequestHeader("Authorization", _apiKey);
-                
+                request.SetRequestHeader("source", AlturaUser.GetSource());
 
                 {
                     yield return request.SendWebRequest();
@@ -184,32 +158,34 @@ namespace AlturaNFT
 
                     if (request.error != null)
                     {
+                        item = null;
                         if(OnErrorAction!=null)
                             OnErrorAction($"Null data. Response code: {request.responseCode}. Result {jsonResult}");
                         if(debugErrorLog)
-                            Debug.Log($"Null data. Response code: {request.responseCode}. Result {jsonResult}");
+                            Debug.Log($" (⊙.◎) Null data. Response code: {request.responseCode}. Result {jsonResult}");
                         if(afterError!=null)
                             afterError.Invoke();
-                        collectionModel = null;
                         //yield break;
                     }
                     else
                     {
-                        collectionModel = JsonConvert.DeserializeObject<Collection_model>(
+                        //Fill Data Model from recieved class
+                        item = JsonConvert.DeserializeObject<Items_model>(
                             jsonResult,
                             new JsonSerializerSettings
                             {
-                            NullValueHandling = NullValueHandling.Ignore,
-                            MissingMemberHandling = MissingMemberHandling.Ignore
+                                NullValueHandling = NullValueHandling.Ignore,
+                                MissingMemberHandling = MissingMemberHandling.Ignore
                             });
                         
                         if(OnCompleteAction!=null)
-                            OnCompleteAction.Invoke(collectionModel);
+                            OnCompleteAction.Invoke(item);
                         
                         if(afterSuccess!=null)
                             afterSuccess.Invoke();
                         
-                            Debug.Log($"Get Many Collection Models" );
+                        if(debugErrorLog)
+                            Debug.Log("Success , view NFTs model" );
                     }
                 }
                 request.Dispose();
